Remove deprecated crypto code
