+++ crypto/src/main/java/org/cryptimeleon/incentive/crypto/IncentiveSystem.java
@@ -114,11 +114,6 @@ public class IncentiveSystem {
         return Setup.storeKeyGen();
     }
 
-    @Deprecated
-    public PromotionParameters legacyPromotionParameters() {
-        return new PromotionParameters(BigInteger.ONE, 1);
-    }
-
     /**
      * Issue a ECDSA signature on a {@code (upk, id-info)} tuple and return an object containing all data, the signature
      * und the verification key.
@@ -599,143 +594,6 @@ public class IncentiveSystem {
         );
     }
 
-
-    /*
-     * implementation of the Credit {@literal <}-{@literal >}Earn protocol
-     */
-
-    /**
-     * Generate an earn request that blinds the token and signature such that the provider can compute a signature on
-     * a matching token with added value.
-     *
-     * @param token             the token to update
-     * @param providerPublicKey the public key of the provider
-     * @param userKeyPair       the key pair of the user submitting the request
-     * @return request to give to a provider
-     */
-    @Deprecated
-    public EarnRequest generateEarnRequest(Token token, ProviderPublicKey providerPublicKey, UserKeyPair userKeyPair) {
-        // Compute pseudorandom value from the token that is used to blind the commitment
-        // This makes this algorithm deterministic
-        var s = pp.getPrfToZn().hashThenPrfToZn(userKeyPair.getSk().getPrfKey(), token, "CreditEarn");
-
-        // Blind commitments and change representation of signature such that it is valid for blinded commitments
-        // The blinded commitments and signature are sent to the provider
-        return new EarnRequest(
-                (SPSEQSignature) pp.getSpsEq().chgRep(
-                        token.getSignature(),
-                        s,
-                        providerPublicKey.getPkSpsEq()
-                ),
-                token.getCommitment0().pow(s).compute(),  // Compute for concurrent computation
-                pp.getG1Generator().pow(s).compute()
-        );
-    }
-
-    /**
-     * Generate the response for users' earn requests to update the blinded token.
-     * Computes a valid signature on a blinded token matching the received token, but with k more points.
-     *
-     * @param earnRequest     the earn request to process. It can be assumed, that all group elements of earnRequest
-     *                        are already computed since they are usually directly parsed from a representation
-     * @param providerKeyPair the provider key pair
-     * @return a signature on a blinded, updated token
-     */
-    @Deprecated
-    public SPSEQSignature generateEarnRequestResponse(PromotionParameters promotionParameters, EarnRequest earnRequest, Vector<BigInteger> deltaK, ProviderKeyPair providerKeyPair) {
-        var C0 = earnRequest.getC0();
-        var C1 = earnRequest.getC1();
-
-        // Verify the blinded signature for the blinded commitment is valid
-        var isSignatureValid = pp.getSpsEq().verify(
-                providerKeyPair.getPk().getPkSpsEq(),
-                earnRequest.getBlindedSignature(),
-                C0,
-                C1,
-                C1.pow(promotionParameters.getPromotionId())
-        );
-        if (!isSignatureValid) {
-            throw new IllegalArgumentException("Signature is not valid");
-        }
-
-        // Sign a blinded commitment with k more points
-
-        var Q = providerKeyPair.getSk().getTokenPointsQ(promotionParameters);
-        var K = deltaK.map(k -> pp.getBg().getG1().getZn().createZnElement(k));
-
-        return (SPSEQSignature) pp.getSpsEq().sign(
-                providerKeyPair.getSk().getSkSpsEq(),
-                C0.op(C1.pow(Q.innerProduct(K))).compute(), // Add k blinded point to the commitment
-                C1,
-                C1.pow(promotionParameters.getPromotionId())
-        );
-    }
-
-    /**
-     * @param earnRequest       the earn request that was originally sent. Again, it can be assumed that all elements
-     *                          are computed since the earnRequest usually is serialized to a representation before.
-     * @param changedSignature  the signature computed by the provider
-     * @param deltaK            the increase for the users token value
-     * @param token             the old token
-     * @param providerPublicKey public key of the provider
-     * @param userKeyPair       keypair of the user
-     * @return new token with value of the old token + k
-     */
-    @Deprecated
-    public Token handleEarnRequestResponse(PromotionParameters promotionParameters,
-                                           EarnRequest earnRequest,
-                                           SPSEQSignature changedSignature,
-                                           Vector<BigInteger> deltaK,
-                                           Token token,
-                                           ProviderPublicKey providerPublicKey,
-                                           UserKeyPair userKeyPair) {
-
-        // Pseudorandom randomness s used for blinding in the request
-        var s = pp.getPrfToZn().hashThenPrfToZn(userKeyPair.getSk().getPrfKey(), token, "CreditEarn");
-        var K = RingElementVector.fromStream(deltaK.stream().map(e -> pp.getBg().getZn().createZnElement(e)));
-
-        // Recover blinded commitments (to match the commitments signed by the prover) with updated value
-        var blindedNewC0 = earnRequest.getC0()
-                .op(providerPublicKey.getTokenPointsH(promotionParameters)
-                        .pow(s)
-                        .innerProduct(K)
-                ).compute();
-        var blindedNewC1 = earnRequest.getC1();
-
-        // Verify signature on recovered commitments
-        var signatureValid = pp.getSpsEq().verify(providerPublicKey.getPkSpsEq(),
-                changedSignature,
-                blindedNewC0,
-                blindedNewC1,
-                blindedNewC1.pow(promotionParameters.getPromotionId()));
-        if (!signatureValid) {
-            throw new IllegalArgumentException("Signature is not valid");
-        }
-
-        // Change representation of signature such that it is valid for un-blinded commitment
-        var newSignature = pp.getSpsEq().chgRep(changedSignature, s.inv(), providerPublicKey.getPkSpsEq());
-
-        // Assemble new token
-        return new Token(
-                blindedNewC0.pow(s.inv()).compute(), // Un-blind commitment
-                token.getDoubleSpendingId(),
-                token.getDoubleSpendRandomness(),
-                token.getZ(),
-                token.getT(),
-                token.getPromotionId(),
-                new RingElementVector(token.getPoints().zip(
-                        K,
-                        RingElement::add
-                )),
-                (SPSEQSignature) newSignature
-        );
-    }
-
-    /*
-     * end of the implementation of the Credit {@literal <}-{@literal >}Earn protocol
-     */
-
-
     /*
      * implementation of the Deduct {@literal <}-{@literal >}Spend protocol
      */
@@ -1033,197 +891,6 @@ public class IncentiveSystem {
         );
     }
 
-
-    /**
-     * Generates a request to add value k to token.
-     *
-     * @param promotionParameters the current promotion's parameters
-     * @param token               the token
-     * @param providerPublicKey   public key of the provider
-     * @param newPoints           the new points vector the token should have
-     * @param userKeyPair         keypair of the user that owns the token
-     * @param tid                 transaction ID, provided by the provider
-     * @param spendDeductTree     the zero knowledge proof for this promotion
-     * @return serializable spendRequest that can be sent to the provider
-     */
-    @Deprecated
-    public SpendRequest generateSpendRequest(PromotionParameters promotionParameters,
-                                             Token token,
-                                             ProviderPublicKey providerPublicKey,
-                                             Vector<BigInteger> newPoints,
-                                             UserKeyPair userKeyPair,
-                                             ZnElement tid,
-                                             SpendDeductTree spendDeductTree
-    ) {
-        // Some local variables and pre-computations to make the code more readable
-        var zp = pp.getBg().getZn();
-        var usk = userKeyPair.getSk().getUsk();
-        var dsid = token.getDoubleSpendingId();
-        var vectorH = providerPublicKey.getH(this.pp, promotionParameters);
-        var newPointsVector = RingElementVector.fromStream(newPoints.stream().map(e -> pp.getBg().getZn().createZnElement(e)));
-
-
-        /* Compute pseudorandom values */
-        // As in credit-earn, we use the PRF to make the algorithm deterministic
-        var R = computeSpendDeductRandomness(userKeyPair.getSk(), token);
-
-        // Prepare a new commitment (cPre0, cPre1) based on the pseudorandom values
-        var exponents = new RingElementVector(R.tS, usk, R.dsidUserS, R.dsrndS, R.zS).concatenate(newPointsVector);
-        var cPre0 = vectorH.innerProduct(exponents).pow(R.uS).compute();
-        var cPre1 = pp.getG1Generator().pow(R.uS).compute();
-
-        /* Enable double-spending-protection by forcing usk and esk becoming public in that case
-           If token is used twice in two different transactions, the provider observes (c0,c1), (c0',c1') with gamma!=gamma'
-           Hence, the provider can easily retrieve usk and esk (using the Schnorr-trick, computing (c0-c0')/(gamma-gamma') for usk, analogously for esk). */
-        // using tid as user choice TODO change this once user choice generation is properly implemented, see issue 75
-        var gamma = Util.hashGammaOld(zp, dsid, tid, cPre0, cPre1, tid);
-        var c = usk.mul(gamma).add(token.getDoubleSpendRandomness());
-
-        /* Compute El-Gamal encryption of esk^*_usr using under secret key esk
-           This allows the provider to decrypt usk^*_usr in case of double spending with the leaked esk.
-           By additionally storing esk^*_prov, the provider can retrieve esk^* and thus iteratively decrypt the new esks. */
-
-        /* Build non-interactive (Fiat-Shamir transformed) ZKP to ensure that the user follows the rules of the protocol */
-        var spendDeductZkp = new SpendDeductBooleanZkp(spendDeductTree, pp, promotionParameters, providerPublicKey);
-        var fiatShamirProofSystem = new FiatShamirProofSystem(spendDeductZkp);
-        var witness = new SpendDeductZkpWitnessInput(usk, token.getZ(), R.zS, token.getT(), R.tS, R.uS, R.dsidUserS, token.getDoubleSpendRandomness(), R.dsrndS, token.getPoints(), newPointsVector);
-        var commonInput = new SpendDeductZkpCommonInput(gamma, c, dsid, cPre0, cPre1, token.getCommitment0());
-        var proof = fiatShamirProofSystem.createProof(commonInput, witness);
-
-        // Assemble request
-        return new SpendRequest(dsid, proof, c, cPre0, cPre1, token.getCommitment0(), token.getSignature());
-    }
-
-    /**
-     * React to a legitimate spend request to allow the user retrieving an updated token with the value decreased by k.
-     * Returns additional data for double-spending protection.
-     *
-     * @param spendRequest        the user's request
-     * @param promotionParameters specifying the promotion that the user wants to spend his points on (i.e. her claim)
-     * @param providerKeyPair     keypair of the provider
-     * @param tid                 transaction id, should be verified by the provider
-     * @param spendDeductTree     the zero knowledge proof to verify for this promotion
-     * @param userChoice          byte representation of the user choice,
-     *                            influences the computation of challenge generator gamma and provider share esk_prov
-     * @return tuple of response to send to the user and information required for double-spending protection
-     */
-    @Deprecated
-    public DeductOutput generateSpendRequestResponse(PromotionParameters promotionParameters,
-                                                     SpendRequest spendRequest,
-                                                     ProviderKeyPair providerKeyPair,
-                                                     ZnElement tid,
-                                                     SpendDeductTree spendDeductTree,
-                                                     UniqueByteRepresentable userChoice) {
-
-        /* Verify that the request is valid and well-formed */
-
-        // Verify signature of the old token (C1 must be g1 according to ZKP in T2 paper. We omit the ZKP and use g1 instead of C1)
-        var signatureValid = pp.getSpsEq().verify(providerKeyPair.getPk().getPkSpsEq(),
-                spendRequest.getSigma(),
-                spendRequest.getCommitmentC0(),
-                pp.getG1Generator(),
-                pp.getG1Generator().pow(promotionParameters.getPromotionId()) // Verify token from correct promotion
-        );
-        if (!signatureValid) {
-            throw new IllegalArgumentException("Signature of the request is not valid!");
-        }
-
-        // Validate ZKP
-        var spendDeductZkp = new SpendDeductBooleanZkp(spendDeductTree, pp, promotionParameters, providerKeyPair.getPk());
-        var fiatShamirProofSystem = new FiatShamirProofSystem(spendDeductZkp);
-        // using tid as user choice TODO change this once user choice generation is properly implemented, see issue 75
-        var gamma = Util.hashGammaOld(pp.getBg().getZn(), spendRequest.getDsid(), tid, spendRequest.getCPre0(), spendRequest.getCPre1(), userChoice);
-        var commonInput = new SpendDeductZkpCommonInput(spendRequest, gamma);
-        var proofValid = fiatShamirProofSystem.checkProof(commonInput, spendRequest.getSpendDeductZkp());
-        if (!proofValid) {
-            throw new IllegalArgumentException("ZKP of the request is not valid!");
-        }
-
-        /* Request is valid. Compute new blinded token and signature
-         *
-         * Retrieve esk*_prov via PRF.
-         * Need to use double-spending ID of the spent token as well as the transaction ID in addition to preliminary commitment as input for the PRF
-         * to ensure that different spendings of the same token lead to different esk_prov.
-         *
-         * Also need to include object modelling the reward that the user chose to claim.
-         * This is done by including a PRFtoZn image hashedClaim of the hashed spend-deduct tree in the spend request.
-         */
-        var preimage = new ByteArrayAccumulator();
-        preimage.escapeAndSeparate(commonInput.c0Pre);
-        preimage.escapeAndSeparate(commonInput.c1Pre);
-        preimage.escapeAndSeparate(spendRequest.getDsid());
-        preimage.escapeAndSeparate(tid);
-        preimage.escapeAndSeparate(userChoice);
-        var eskStarProv = pp.getPrfToZn().hashThenPrfToZn(providerKeyPair.getSk().getBetaProv(), new ByteArrayImplementation(preimage.extractBytes()), "eskStarProv");
-
-        // Compute blind signature on new, still blinded commitment
-        var cPre0 = spendRequest.getCPre0();
-        var cPre1 = spendRequest.getCPre1();
-        var sigmaPrime = (SPSEQSignature) pp.getSpsEq().sign(
-                providerKeyPair.getSk().getSkSpsEq(),
-                cPre0.op(cPre1.pow(eskStarProv.mul(providerKeyPair.getSk().getQ().get(1)))),
-                cPre1,
-                cPre1.pow(promotionParameters.getPromotionId())
-        );
-
-        // Assemble providers and users output and return as a tuple
-        return new DeductOutput(
-                new SpendResponse(sigmaPrime, eskStarProv),
-                new DoubleSpendingTag(commonInput.c, gamma)
-        );
-    }
-
-    /**
-     * Process the response to a spendRequest and compute the updated token.
-     *
-     * @param spendResponse     the provider's response
-     * @param spendRequest      the original request
-     * @param token             the old token
-     * @param providerPublicKey public key of the provider
-     * @param userKeyPair       keypair of the user
-     * @return token with the value of the old token + k
-     */
-    @Deprecated
-    public Token handleSpendRequestResponse(PromotionParameters promotionParameters,
-                                            SpendResponse spendResponse,
-                                            SpendRequest spendRequest,
-                                            Token token,
-                                            Vector<BigInteger> newPoints,
-                                            ProviderPublicKey providerPublicKey,
-                                            UserKeyPair userKeyPair
-    ) {
-
-        var newPointsVector = RingElementVector.fromStream(newPoints.stream().map(e -> pp.getBg().getZn().createZnElement(e)));
-
-        // Re-compute pseudorandom values
-        var R = computeSpendDeductRandomness(userKeyPair.getSk(), token);
-
-        // Verify the signature on the new, blinded commitment
-        var blindedCStar0 = spendRequest.getCPre0().op(providerPublicKey.getH().get(1).pow(spendResponse.getEskProvStar().mul(R.uS)));
-        var blindedCStar1 = pp.getG1Generator().pow(R.uS); // Recompute, just to make sure
-        var valid = pp.getSpsEq().verify(providerPublicKey.getPkSpsEq(),
-                spendResponse.getSigma(),
-                blindedCStar0,
-                blindedCStar1,
-                blindedCStar1.pow(promotionParameters.getPromotionId()));
-        if (!valid) {
-            throw new IllegalArgumentException("Signature is not valid");
-        }
-
-        // Build new token
-        return new Token(
-                blindedCStar0.pow(R.uS.inv()), // Unblind commitment
-                // Same as unblinded CStar1
-                R.dsidUserS.add(spendResponse.getEskProvStar()),
-                R.dsrndS,
-                R.zS,
-                R.tS,
-                token.getPromotionId(),
-                new RingElementVector(newPointsVector),
-                // Change representation of signature to match the un-blinded commitments
-                (SPSEQSignature) pp.getSpsEq().chgRep(spendResponse.getSigma(), R.uS.inv(), providerPublicKey.getPkSpsEq())
-        );
-    }
 
     /*
      * end of the implementation of the Deduct {@literal <}-{@literal >}Spend protocol
